Restore decimal value to int conversion

+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/internal/values/DecimalValue.java
@@ -189,7 +189,7 @@ public class DecimalValue implements SimpleValue, BDecimal {
             throw ErrorUtils.createNumericConversionError(this.stringValue(null), PredefinedTypes.TYPE_DECIMAL,
                                                           PredefinedTypes.TYPE_INT);
         }
-        return (long) Math.rint(value.longValue());
+        return (long) Math.rint(value.doubleValue());
     }
 
     /**
